# Optimize NullableSingleInputAggregationFunction when entire block is null