--- a/pinot-core/src/main/java/org/apache/pinot/core/query/aggregation/function/NullableSingleInputAggregationFunction.java
+++ b/pinot-core/src/main/java/org/apache/pinot/core/query/aggregation/function/NullableSingleInputAggregationFunction.java
@@ -81,7 +81,9 @@ public abstract class NullableSingleInputAggregationFunction<I, F extends Compar
       return;
     }
 
-    forEachNotNull(length, roaringBitmap.getIntIterator(), consumer);
+    if (roaringBitmap.getCardinality() < length) {
+      forEachNotNull(length, roaringBitmap.getIntIterator(), consumer);
+    }
   }
 
   /**
@@ -119,6 +121,11 @@ public abstract class NullableSingleInputAggregationFunction<I, F extends Compar
    */
   public <A> A foldNotNull(int length, @Nullable RoaringBitmap roaringBitmap, A initialAcum, Reducer<A> reducer) {
     IntIterator intIterator = roaringBitmap == null ? null : roaringBitmap.getIntIterator();
+
+    if (_nullHandlingEnabled && roaringBitmap != null && roaringBitmap.getCardinality() >= length) {
+      return initialAcum;
+    }
+
     return foldNotNull(length, intIterator, initialAcum, reducer);
   }