--- a/pinot-perf/src/main/java/org/apache/pinot/perf/BenchmarkModeAggregation.java
+++ b/pinot-perf/src/main/java/org/apache/pinot/perf/BenchmarkModeAggregation.java
@@ -58,11 +58,11 @@ public class BenchmarkModeAggregation extends AbstractAggregationFunctionBenchma
   @Param({"100", "50", "0"})
   public int _nullHandlingEnabledPerCent;
   private boolean _nullHandlingEnabled;
-  @Param({"2", "4", "8", "16", "32", "64", "128"})
+  @Param({"1", "2", "4", "8", "16", "32", "64", "128"})
   protected int _nullPeriod;
   private final Random _segmentNullRandomGenerator = new Random(42);
-  private double _modeIgnoringNull;
-  private double _modeNullAware;
+  private Double _modeIgnoringNull;
+  private Double _modeNullAware;
   private final int _numDocs = DocIdSetPlanNode.MAX_DOC_PER_CALL;
 
   public static void main(String[] args)
@@ -122,11 +122,15 @@ public class BenchmarkModeAggregation extends AbstractAggregationFunctionBenchma
         .get()
         .getKey()
         .doubleValue();
-    _modeNullAware = nullAwareDistribution.entrySet().stream()
-        .max(Comparator.comparingInt(Map.Entry::getValue))
-        .get()
-        .getKey()
-        .doubleValue();
+    if (nullAwareDistribution.isEmpty()) {
+      _modeNullAware = null;
+    } else {
+      _modeNullAware = nullAwareDistribution.entrySet().stream()
+              .max(Comparator.comparingInt(Map.Entry::getValue))
+              .get()
+              .getKey()
+              .doubleValue();
+    }
   }
 
   @Override