- Currently, if null handling is enabled and any aggregation function extending `NullableSingleInputAggregationFunction` encounters a block where every value is `null`, it will unnecessarily iterate through every entry in the null bitmap with a check to see whether any non-null values were encountered in each iteration.
- Instead, we can skip over this completely based on the null bitmap's cardinality. This optimization already exists in some of the other aggregation functions that don't currently extend `NullableSingleInputAggregationFunction` - 

https://github.com/apache/pinot/blob/0f2dcb770dca414b02dc4d495d4b4458fa2e91ec/pinot-core/src/main/java/org/apache/pinot/core/query/aggregation/function/SumAggregationFunction.java#L135

https://github.com/apache/pinot/blob/0f2dcb770dca414b02dc4d495d4b4458fa2e91ec/pinot-core/src/main/java/org/apache/pinot/core/query/aggregation/function/AvgAggregationFunction.java#L102
- Some of the existing unit tests such as `ModeAggregationFunctionTest` and `AbstractPercentileAggregationFunctionTest` already provide coverage for this scenario.