@richardstartin 2024-08-06
To extend this idea, it would be better to check if the bitmap is empty with the isEmpty method (constant time) rather than compute its cardinality (worst case linear time).