@Jackie-Jiang 2024-08-06
@richardstartin Good point! Is there an API to check if the bitmap is all set? Ideally we want to early terminate whenever there is an unset bit