@gortiz 2024-08-07
> Ah thanks, I hadn't looked too closely at the implementation and didn't realize that it could be worst case linear time. In that case, I'm not sure if this optimization makes much sense.

Can you use my branch with the benchmarks to compare performance?