@yashmayya 2024-08-08
I've updated the mode aggregation benchmark and pushed the change to this branch itself. Here are the results:

Without this optimization:
```
Benchmark                      (_nullHandlingEnabledPerCent)  (_nullPeriod)   Mode  Cnt   Score   Error   Units
BenchmarkModeAggregation.test                            100              1  thrpt   50  46.303 ± 0.274  ops/ms
BenchmarkModeAggregation.test                            100              2  thrpt   50  36.063 ± 0.203  ops/ms
BenchmarkModeAggregation.test                            100              4  thrpt   50  31.343 ± 0.128  ops/ms
BenchmarkModeAggregation.test                            100              8  thrpt   50  30.965 ± 0.631  ops/ms
BenchmarkModeAggregation.test                            100             16  thrpt   50  31.089 ± 0.154  ops/ms
BenchmarkModeAggregation.test                            100             32  thrpt   50  32.219 ± 0.192  ops/ms
BenchmarkModeAggregation.test                            100             64  thrpt   50  32.074 ± 0.157  ops/ms
BenchmarkModeAggregation.test                            100            128  thrpt   50  32.154 ± 0.388  ops/ms
BenchmarkModeAggregation.test                             50              1  thrpt   50  47.861 ± 8.267  ops/ms
BenchmarkModeAggregation.test                             50              2  thrpt   50  34.433 ± 0.933  ops/ms
BenchmarkModeAggregation.test                             50              4  thrpt   50  31.789 ± 0.204  ops/ms
BenchmarkModeAggregation.test                             50              8  thrpt   50  32.006 ± 0.303  ops/ms
BenchmarkModeAggregation.test                             50             16  thrpt   50  32.274 ± 0.243  ops/ms
BenchmarkModeAggregation.test                             50             32  thrpt   50  32.425 ± 0.200  ops/ms
BenchmarkModeAggregation.test                             50             64  thrpt   50  32.301 ± 0.171  ops/ms
BenchmarkModeAggregation.test                             50            128  thrpt   50  32.473 ± 0.147  ops/ms
BenchmarkModeAggregation.test                              0              1  thrpt   50  28.122 ± 0.310  ops/ms
BenchmarkModeAggregation.test                              0              2  thrpt   50  27.857 ± 0.163  ops/ms
BenchmarkModeAggregation.test                              0              4  thrpt   50  27.667 ± 0.164  ops/ms
BenchmarkModeAggregation.test                              0              8  thrpt   50  27.933 ± 0.188  ops/ms
BenchmarkModeAggregation.test                              0             16  thrpt   50  27.772 ± 0.159  ops/ms
BenchmarkModeAggregation.test                              0             32  thrpt   50  28.018 ± 0.187  ops/ms
BenchmarkModeAggregation.test                              0             64  thrpt   50  28.060 ± 0.358  ops/ms
BenchmarkModeAggregation.test                              0            128  thrpt   50  28.159 ± 0.192  ops/ms
```

With this optimization:
```
Benchmark                      (_nullHandlingEnabledPerCent)  (_nullPeriod)   Mode  Cnt       Score       Error   Units
BenchmarkModeAggregation.test                            100              1  thrpt   50  154371.180 ±   592.954  ops/ms
BenchmarkModeAggregation.test                            100              2  thrpt   50      36.142 ±     0.301  ops/ms
BenchmarkModeAggregation.test                            100              4  thrpt   50      32.071 ±     0.403  ops/ms
BenchmarkModeAggregation.test                            100              8  thrpt   50      31.035 ±     0.396  ops/ms
BenchmarkModeAggregation.test                            100             16  thrpt   50      31.827 ±     0.201  ops/ms
BenchmarkModeAggregation.test                            100             32  thrpt   50      32.297 ±     0.138  ops/ms
BenchmarkModeAggregation.test                            100             64  thrpt   50      32.251 ±     0.158  ops/ms
BenchmarkModeAggregation.test                            100            128  thrpt   50      32.452 ±     0.161  ops/ms
BenchmarkModeAggregation.test                             50              1  thrpt   50   67106.709 ± 37843.797  ops/ms
BenchmarkModeAggregation.test                             50              2  thrpt   50      34.638 ±     0.855  ops/ms
BenchmarkModeAggregation.test                             50              4  thrpt   50      32.408 ±     0.210  ops/ms
BenchmarkModeAggregation.test                             50              8  thrpt   50      29.867 ±     0.635  ops/ms
BenchmarkModeAggregation.test                             50             16  thrpt   50      31.973 ±     0.408  ops/ms
BenchmarkModeAggregation.test                             50             32  thrpt   50      32.236 ±     0.293  ops/ms
BenchmarkModeAggregation.test                             50             64  thrpt   50      31.478 ±     0.494  ops/ms
BenchmarkModeAggregation.test                             50            128  thrpt   50      32.253 ±     0.276  ops/ms
BenchmarkModeAggregation.test                              0              1  thrpt   50      27.938 ±     0.179  ops/ms
BenchmarkModeAggregation.test                              0              2  thrpt   50      27.863 ±     0.204  ops/ms
BenchmarkModeAggregation.test                              0              4  thrpt   50      28.036 ±     0.202  ops/ms
BenchmarkModeAggregation.test                              0              8  thrpt   50      28.203 ±     0.372  ops/ms
BenchmarkModeAggregation.test                              0             16  thrpt   50      28.106 ±     0.146  ops/ms
BenchmarkModeAggregation.test                              0             32  thrpt   50      28.663 ±     0.352  ops/ms
BenchmarkModeAggregation.test                              0             64  thrpt   50      28.048 ±     0.145  ops/ms
BenchmarkModeAggregation.test                              0            128  thrpt   50      28.103 ±     0.164  ops/ms
```

Also, I tried this optimization with a couple of other aggregation functions and benchmarks locally and saw similar result.